refactor: ssr dev server path
